// star-ui/src/proto/star/v1/telemetry.ts
export enum ConnectionStatus {
    /**
     * Unknown connection state.
     *
     * @generated from protobuf enum value: CONNECTION_STATUS_UNKNOWN = 0;
     */
    UNKNOWN = 0,
    /**
     * Connected and communicating.
     *
     * @generated from protobuf enum value: CONNECTION_STATUS_CONNECTED = 1;
     */
    CONNECTED = 1,
    /**
     * Not connected.
     *
     * @generated from protobuf enum value: CONNECTION_STATUS_DISCONNECTED = 2;
     */
    DISCONNECTED = 2,
    /**
     * Connection in progress.
     *
     * @generated from protobuf enum value: CONNECTION_STATUS_CONNECTING = 3;
     */
    CONNECTING = 3,
    /**
     * Connection error occurred.
     *
     * @generated from protobuf enum value: CONNECTION_STATUS_ERROR = 4;
     */
    ERROR = 4
}
/**
 * Robot operating mode enumeration.
 *
 * @generated from protobuf enum star.v1.RobotMode
 */
export enum RobotMode {
    /**
     * Unknown operating mode.
     *
     * @generated from protobuf enum value: ROBOT_MODE_UNKNOWN = 0;
     */
    UNKNOWN = 0,
    /**
     * Robot is idle (motors off).
     *
     * @generated from protobuf enum value: ROBOT_MODE_IDLE = 1;
     */
    IDLE = 1,
    /**
     * Manual control mode (joystick/UI).
     *
     * @generated from protobuf enum value: ROBOT_MODE_MANUAL = 2;
     */
    MANUAL = 2,
    /**
     * Autonomous navigation mode (Nav2).
     *
     * @generated from protobuf enum value: ROBOT_MODE_AUTONOMOUS = 3;
     */
    AUTONOMOUS = 3,
    /**
     * SLAM mapping mode.
     *
     * @generated from protobuf enum value: ROBOT_MODE_MAPPING = 4;
     */
    MAPPING = 4,
    /**
     * Emergency stop active.
     *
     * @generated from protobuf enum value: ROBOT_MODE_EMERGENCY_STOP = 5;
     */
    EMERGENCY_STOP = 5
}